pyproject.toml
[project]
name = "mcp-na-pratica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.96.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "wikipedia>=1.4.0",
]
